requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.4
defusedxml==0.8.0rc2
Django==4.0
django-allauth==0.61.1
django-appconf==1.0.5
django-compressor==4.4
django-debug-toolbar==4.3.0
django-filter==23.3
django-mail-templated==2.6.5
django-ranged-response==0.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
flake8==7.0.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
Markdown==3.5.1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==10.0.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rpds-py==0.17.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
zipp==3.17.0